moved all Property use cases into separate controllers
